+++ CHANGELOG.md
@@ -3,7 +3,9 @@ Change Log
 
 ## Version 1.0.0-dev08 (baseline: NCM 9.0.80)
 
-- Improve Top Lists from draft
+- Improve Top Lists which was just draft
+- Fix playlist info overlapping with toolbar
+- Fix: Add missing click listeners for toolbar navigation icons
 
 ## Version 1.0.0-dev07 (baseline: NCM 9.0.80)